# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.LendingInfoResponse:
    properties:
      currentInterestRate:
        type: string
      interestEarned:
        type: string
      totalDeposited:
        type: string
    type: object
...
  dto.MarketOverviewResponse:
    properties:
      activePositions:
        type: integer
      activeUsers:
        type: integer
      averageBorrowingAPY:
        type: string
      averageLendingAPY:
        type: string
      totalBorrowed:
        type: string
      totalValueLocked:
        type: string
    type: object
  dto.TokenMarketData:
    properties:
      availableLiquidity:
        type: string
      borrowingAPY:
        type: string
      collateralFactor:
        type: string
      lendingAPY:
        type: string
      token:
        $ref: '#/definitions/dto.TokenMetadata'
      totalBorrowed:
        type: string
      totalDeposited:
        type: string
      totalSupply:
        type: string
    type: object
  dto.TokenMetadata:
    properties:
      address:
        type: string
      decimals:
        type: integer
      logoURI:
        type: string
      name:
        type: string
      priceUSD:
        type: string
      symbol:
        type: string
    type: object
  dto.TokensMarketResponse:
    properties:
      tokens:
        items:
          $ref: '#/definitions/dto.TokenMarketData'
        type: array
    type: object
...
  dto.TransactionListResponse:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPage:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/dto.TransactionResponse'
        type: array
    type: object
  dto.TransactionRequest:
    properties:
      amount:
        type: string
    required:
    - amount
    type: object
  dto.TransactionResponse:
    properties:
      amount:
        type: string
      blockNumber:
        type: integer
      createdAt:
        type: string
      gasPrice:
        type: string
      gasUsed:
        type: integer
      hash:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/dto.TransactionStatus'
      tokenAddress:
        type: string
      type:
        $ref: '#/definitions/dto.TransactionType'
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  dto.TransactionStatus:
    enum:
    - pending
    - confirmed
    - failed
    type: string
    x-enum-varnames:
    - TransactionStatusPending
    - TransactionStatusConfirmed
    - TransactionStatusFailed
  dto.TransactionType:
    enum:
    - deposit
    - withdraw
    - borrow
    - repay
    - liquidate
    type: string
    x-enum-varnames:
    - TransactionTypeDeposit
    - TransactionTypeWithdraw
    - TransactionTypeBorrow
    - TransactionTypeRepay
    - TransactionTypeLiquidate
...
host: localhost:8080
info:
  contact: {}
  description: API for decentralized lending and borrowing platform
  title: Lending & Borrowing Platform API
  version: "1.0"
paths:
...
  /lending/balance:
    get:
      consumes:
      - application/json
      description: Get user's current balance in the lending pool
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get lending balance
      tags:
      - lending
  /lending/deposit:
    post:
      consumes:
      - application/json
      description: Make a deposit into the lending pool to earn interest
      parameters:
      - description: Deposit amount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deposit tokens to lending pool
      tags:
      - lending
  /lending/info:
    get:
      consumes:
      - application/json
      description: Get detailed lending information for the user including interest
        earned
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LendingInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get lending information
      tags:
      - lending
  /lending/pool-info:
    get:
      consumes:
      - application/json
      description: Get information about the lending pool including total deposited
        and interest rate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get lending pool information
      tags:
      - lending
  /lending/transactions:
    get:
      consumes:
      - application/json
      description: Get paginated history of user's lending transactions
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TransactionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get lending transaction history
      tags:
      - lending
  /lending/withdraw:
    post:
      consumes:
      - application/json
      description: Withdraw deposited tokens from the lending pool
      parameters:
      - description: Withdraw amount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Withdraw tokens from lending pool
      tags:
      - lending
...
  /market/overview:
    get:
      consumes:
      - application/json
      description: Get overview of the market including TVL, borrowed amounts, and
        rates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MarketOverviewResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get market overview
      tags:
      - market
  /market/tokens:
    get:
      consumes:
      - application/json
      description: Get detailed market data for all supported tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokensMarketResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get tokens market data
      tags:
      - market
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"